Move validation calls out of commandExecutor
